Updates for RN0.29 compatibility
Change Manager to Module, extend ReactContextBaseJavaModule and use
getCurrentActivity to get a reference to the activity, instead of having it pass
through the package's constructor.

Change API of component - text for task description now provided as label prop
instead of child Text component.

Update installation and usage in README.md.